requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
Markdown==3.3.7
psycopg2==2.9.3
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
